Got multiline text working by using arrays.
Newlines are a pain. I should add on to the library and add a class that will automatically handle newlines. It would seperate the string into an array, and then position everything so that newlines and wrapped lines are taken into account. Text stuff is the most annoying work I've done programming so far.